wysylanie polecen

+++ main/fast_add.h
@@ -13,13 +13,13 @@ typedef struct
 {
 	uint32_t counter;
 	uint32_t delay;
-	uint8_t * value;
-	uint8_t max, min, sign;
-	void (*func)(uint8_t);
+	uint32_t * value;
+	uint32_t max, min, sign;
+	void (*func)(uint32_t);
 }fast_add_t;
 
-void fastProcessStart(uint8_t * value, uint8_t max, uint8_t min, fast_process_sign sign, void (*func)(uint8_t));
-void fastProcessStop(uint8_t * value);
+void fastProcessStart(uint32_t * value, uint32_t max, uint32_t min, fast_process_sign sign,  void (*func)(uint32_t));
+void fastProcessStop(uint32_t * value);
 void fastProcessDeInit(void);
 void fastProcessStartTask(void);
